Allows setting of the --continue flag with flog
This is useful because the version of ruby_parser in use by flog does
not support ruby 2.0's keyword arguments. Passing the --continue flag
allows metric_fu's flog to run to completion. Then saikuro barfs, but
hey, at least flog runs.

This can then be set in the .metrics file with:

    MetricFu::Configuration.run do |config|
      config.flog = config.flog.merge({continue: true})
    end