Call RailsBestPractices directly to get access to the array of errors.
We now have the URL, so we can provide a link to rails-best-pracitces.com
By calling the lib directly we can avoid IO overhead, however we want to
be sure that the RailsBestPractices maintains the interface we are
expecting, since it's not a publicly advertised interface.
I have added some specs that will error if the interface to
RailsBestPractices changes.

huh